requirements/base-requirements.txt
Django>=2.2.8
djangorestframework>=3.10.3
bleach>=3.1.0
